requirements.txt
pywebview
argh
requests
tqdm
sqlite-database
pyyaml